Moved AcmeCache into its own module
Renamed `fs` to `cache` and moved AcmeCache into it.

Also removed the macro implementations -- they were unnecessary after
settling on forcing a reference to be passed into `order()`.
